Breaking all sorts of tests, hunting for integer conversion bugs.

+++ src/fen.c
@@ -55,7 +55,7 @@ static bool parseToMove(char * t, int* result) {
     return true;
 }
 
-static bool parseFiftyMove(char* t, int* result) {
+static bool parseFiftyMove(char* t, long* result) {
     char* endToken;
     long parsed = strtol(t, &endToken, 10);
 
@@ -81,7 +81,7 @@ static bool parseFiftyMove(char* t, int* result) {
 }
 
 static bool parseEpFile(char* t, int* result) {
-    int len = strlen(t);
+    unsigned long len = strlen(t);
     bool valid = true;
 
     if(len == 1 && strcmp("-", t) == 0) {
@@ -212,7 +212,7 @@ int parseFen(GameState* state, char* fenStr) {
     int toMove = COLOR_WHITE;
     int castleFlags = 0;
     int epFile = NO_EP_FILE;
-    int fiftyMove = 0;
+    long fiftyMove = 0;
 
     tokenBuffer = createTokenBuffer(_FEN_MAX_TOKENS, _FEN_MAX_TOKEN_LEN);
